move java extraction rules behind the indexer language registry
S3-JAVA (ARCH-012), the second of sixteen one-language rollback units and the
first sibling of S3-KOTLIN. Java's tree-sitter rule file, compiled-rule cache,
member-call callsite marker, and receiver-call resolution engine move out of the
indexer god file into `codestory-indexer/src/languages/java.rs`, behind the
`LanguageExtraction` registry in `languages/mod.rs`. Every language-keyed
dispatch that used to spell "java" now reads the registry first and keeps its
residual `match` for the fourteen languages that have not moved; the last
sibling package deletes those arms.

The move is mechanical: 586 non-blank lines were deleted from `lib.rs` and the
same 586 lines are the function bodies in `languages/java.rs`, byte-identical
apart from renaming the two entry points
(`collect_java_receiver_call_edges` -> `receiver_call_specs`,
`java_member_call` -> `member_call`). Twenty-four functions left `lib.rs` and
twenty-four landed in the module, plus the new `java_language()` parser
constructor. `java_named_child` and `collect_java_declaration_span_overrides`
stayed because declaration spans are a separate cross-language dispatch with no
registry field.

Proof: `tests/language_extraction_snapshot.rs` gains one `SnapshotCase` row that
pins the full rendered projection of both Java fixtures -- the fidelity_lab file
through the store pipeline (resolution included) and the tictactoe file through
raw `index_file`. Both goldens were captured before the production edit and are
byte-identical after it (sha256 17533fdc.. / 31c9750f..). The threshold suites
are not sufficient on their own: changing Java's qualified-name delimiter to
"::" leaves fidelity_regression, tictactoe_language_coverage, and all 269 lib
unit tests green while both new snapshots go red.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. Java
  fidelity_regression and tictactoe_language_coverage projections are proven
  byte-identical before and after the move, and packet_search_eval is unmoved.
- `LanguageRuleset::Java`'s arm in `compiled_rules` is now unreachable and
  returns a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, exactly as the Kotlin arm already is.
- One additive `LanguageCommentProfile` row for "java" in
  `codestory_contracts::language_support`, and the matching deletion of "java"
  from the CLI's residual comment roster in `output.rs::ansi_highlight_line`.
  The marker is "//" on both sides, so rendering is unchanged. The CLI roster
  and the indexer route-comment roster still disagree on cpp/swift/tsx/jsx and
  vue/astro; those are separate facts with separate owners and are left alone.
- `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim tiers,
  `supported_extensions`, `is_structural_source_path`, and every existing
  language claim are identical.
- `semantic::JavaSemanticResolver` stays behind `dedicated_semantic_resolver`
  because the resolver types are private to that module; the registry records
  the choice as `uses_generic_semantic_resolver: false`.
- No version surface touched; everything stays 0.16.3.

Closes #1677
Refs #1243
Refs #1179
